+++ client/src/main/java/io/pravega/client/tables/TableWriter.java
@@ -16,21 +16,30 @@ import java.util.concurrent.CompletableFuture;
  * Defines all operations that can be used to modify a Table.
  *
  * Types of updates:
- *  * * Blind Updates will insert and/or overwrite any existing values for the given Key, regardless of whether that Key
- *  * previously existed or not, and regardless of what that Key's version is.
- *  * * Conditional Updates will only overwrite an existing value if the specified version matches that Key's version. If
- *  * the key does not exist, the {@link TableKey} or {@link TableEntry} must have been created with {@link KeyVersion#NOT_EXISTS}
- *  * in order for the update to succeed.
- *  * * Blind Removals will remove a Key regardless of what that Key's version is. The operation will also succeed (albeit
- *  * with no effect) if the Key does not exist.
- *  * * Conditional Removals will remove a Key only if the specified {@link TableKey#getVersion()} matches that Key's version.
- *  * It will also fail (with no effect) if the Key does not exist and Version is not set to {@link KeyVersion#NOT_EXISTS}.
+ * * Blind Updates will insert and/or overwrite any existing values for the given Key, regardless of whether that Key
+ * previously existed or not, and regardless of what that Key's version is.
+ * * Conditional Updates will only overwrite an existing value if the specified {@link KeyVersion} matches that Key's version.
+ * * Blind Removals will remove a Key regardless of what that Key's version is. The operation will also succeed (albeit
+ * with no effect) if the Key does not exist.
+ * * Conditional Removals will remove a Key only if the specified {@link KeyVersion} matches that Key's version.
  *
  * @param <KeyT> Table Key Type.
  * @param <ValueT> Table Value Type.
  */
 @Beta
 public interface TableWriter<KeyT, ValueT> extends AutoCloseable {
+
+    /**
+     * Inserts a new Key-Value pair in this Table, but only if the Key does not already exist. Consider using
+     * {@link #put(KeyT, ValueT)} should the intent be to insert or update an existing key.
+     *
+     * @param key   The Key to insert or update.
+     * @param value The Value to set for the Key.
+     * @return A CompletableFuture that, when completed, will contain the {@link KeyVersion} associated with the newly
+     * inserted or updated entry.
+     */
+    CompletableFuture<KeyVersion> create(KeyT key, ValueT value);
+
     /**
      * Performs a Blind Update by inserting or updating the Value for a Key in this Table, regardless of whether the Key
      * exists or not or what its Version is.
@@ -43,13 +52,11 @@ public interface TableWriter<KeyT, ValueT> extends AutoCloseable {
     CompletableFuture<KeyVersion> put(KeyT key, ValueT value);
 
     /**
-     * Performs a Conditional Update by inserting or updating the Value for a Key in this Table.
+     * Performs a Conditional Update by updating the Value for an existing Key in this Table.
      *
      * @param key            The Key to insert or update.
      * @param value          The Value to set for the Key.
-     * @param compareVersion A {@link KeyVersion} that will need to match the version of the Key in the Table. Consider
-     *                       using {@link KeyVersion#NOT_EXISTS} if this is used to insert a Key which must not previously
-     *                       exist.
+     * @param compareVersion A {@link KeyVersion} that will need to match the version of the Key in the Table.
      * @return A CompletableFuture that, when completed, will contain the {@link KeyVersion} associated with the newly
      * inserted or updated entry.
      */
