+++ client/src/main/java/io/pravega/client/tables/TableWriterConfig.java
@@ -21,14 +21,14 @@ import lombok.Data;
 public class TableWriterConfig implements Serializable {
 
     private static final long serialVersionUID = 1L;
-    private final int initalBackoffMillis;
+    private final int initialBackoffMillis;
     private final int maxBackoffMillis;
     private final int retryAttempts;
     private final int backoffMultiple;
     private final long transactionTimeoutTime;
 
     public static final class TableWriterConfigBuilder {
-        private int initalBackoffMillis = 1;
+        private int initialBackoffMillis = 1;
         private int maxBackoffMillis = 20000;
         private int retryAttempts = 10;
         private int backoffMultiple = 10;
